+++ styled_text/example/ios/Flutter/flutter_export_environment.sh
@@ -1,11 +1,13 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
+
 export "FLUTTER_ROOT=/Users/surielbe/development/flutter"
 export "FLUTTER_APPLICATION_PATH=/Users/surielbe/flutter-apps/styled_text_package/styled_text/example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
+
 export "FLUTTER_BUILD_NAME=2.0.0"
 export "FLUTTER_BUILD_NUMBER=2.0.0"
 export "DART_OBFUSCATION=false"