+++ styled_text/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 3.0.2
+
+* **[Breaking Changes]** In `StyledTextWidgetBuilderTag.builder` now the tag attributes are passed as the second parameter.
+
+## 3.0.1
+
+* Fixed a bug when unmounting a widget from a tree during text parsing.
+
+## 3.0.0
+
+* Flutter 2.2 compatibility.
+
+## 3.0.0-beta+2
+
+* **[Breaking Changes]** Changed the way text tags are styled. Now, instead of `TextStyle` and classes that implement it, you need to use new classes that describe the styles and behavior of tags.
+* **[Breaking changes]** The `newLineAsBreaks` parameter is now `true` by default.
+* Added the ability to insert any widgets into the text.
+
+## 2.0.0
+
+* Fixed a typo in the example.
+
 ## 2.0.0-nullsafety.0
 
 * Migrate to null safety.